Differentiate TDVP1 algorithm for vectorized state
Add a new tdvp1vec! function that replaces tdvp1! when used on
vectorized states. It is unclear if the measurement can be done while
the sweep is still running, or if it must wait until every site has been
updated. The second option is certainly true, so the new function does
exactly that, performing the measurements only after the sweep has ended.