Move !isempty(measurements) check out of print fns
The printoutput functions, at the beginning, check that the
measurement_ts(bc) vector is not empty before writing its values to the
files. This happens usually during the first sweep, when no measurements
have been performed yet.
However, the test in the "if" block which activates the functions also
refers to such vector, by calling its last element. Of course, if the
vector is empty this call throws a BoundsError.
The test in this "if" block must therefore also check that
measurement_ts(cb) is not empty (and not the function themselves).