Fix use of initial state in printoutput_data
The function needs another keyword argument in the case we want to print
the overlap with the initial state, but this wasn't possible in an
automatical way with the old syntax. This requires an if-block when the
function is called, which (if necessary) adds the initial state to the
kwargs list.