Write basic script for fermionic Markovian closure
The script, based on the one for bosons (called "WSCP..."), performs the
simulation of a system which is practically identical to the original
one, expect for the fact that the modes are fermionic in nature.
For now, almost everything but the tdvp1 algorithm, which was already
done elsewhere, is explicitly written in the script; most of the code
might (and will) be later encapsulated in some functions, so that it is
faster and more readable.