Fix tensor passed to measure_localops! in tdvp1!
We must be careful to pass to the measure_localops! function only the
sites that are have just been updated; this means that with TDVP1 an
expression such as psi[bond+1] fails when the last site of the MPS is
considered (it goes out of bounds).
A different selection is then needed for each different algorithm.