Move on-current-bond check inside measure_localops
This function already knows the operators: they are inside the callback object
The point was that this `ops` argument was used in order to select which
operators were to be measured, namely the ones relative to the bond just updated
by the evolution algorithm.
But this can be done automatically inside this function: it just needs
to know the current site (actually, it already does). By supplying to it
the algorithm, it can then know which bond is the "active" one.