Define "Pos" callback for non-vectorized systems
The old callback object LocalPosMeasurementCallback implicitly used the
formulas for vectorized systems, such as contracting with vecId's in the
measure_localops! function. This doesn't make it possible to use this
object in simulations where the system is represented by a pure state.
Therefore a new callback object, LocalPosVecMeasurementCallback, was
created, with the correct behaviour with vectorized states, while
LocalPosMeasurementCallback was modified so that it can be used with
pure states.