Fix accidental imaginary time step in tdvp1!
Add back to the tdvp1! function the line where the time step given as an
argument is multiplied my the imaginary unit.
This line was removed in an older commit, without paying appropriate
attention to what it did, an as a consequence the time evolution in
tdvp1! happened with an imaginary time when a real time was passed.