Read observables to be measured from JSON file
Instead of hard-coding the observables (and the site they are associated
to) in the main script code, they must now be written in a subdictionary
named "observables" in the JSON parameter file.
This allows the end user to change them as they like without needing to
modify the simulation script.