Fix piecewise definition of thermalized sdf
The previous definition of the thermalized spectral density function
(with chemical potential) involved indicator functions to specify the
domain; however, the expressions multiplied by such indicator functions
were not defined in the whole support of the function.
When Julia tried to evaluate the expression, every term was computed
even if the relative indicator function returned zero, and this produced
some DomainErrors.
This was fixed moving the thermalization procedure in a separate
function with all the necessary if-else conditionals.