Fix wrong sites in dissipation operators
Some sites for the operator in the dissipation part of the master
equation were accidentally given as `1:closure_length` instead of
starting from the first site of the closure. Consequently, the a and
a^dagger operators in the equation did not match: this was what caused
the rapid decrease of the trace.