Use AutoMPO to build Lindblad op terms
Use the AutoMPO syntax as much as possible to build the operators
appearing in the Lindblad operator. The AutoMPO syntax actually accepts
tuples of arbitrary lengths, so it can actually be used to define MPOs
with non trivial terms on more than two sites, such as the Pauli strings
used for the interaction between the pseudomodes and the TEDOPA chain.