Measure trace instead of norm for vectorized MPS
When the MPS represents a vectorized density matrix, it does not make
sense to record its norm at each measurement step: the quantity that
truly represents the normalization of the state is the trace, in this
case; the norm of the MPS corresponds to the purity of the state, which
is interesting but irrelevant in this case.
This addition also requires passing the Vector{Index{...}} of the
physical sites from the tdvp1vec! down to the function that actually
measures the trace, since this information is not included in the MPS
itself.

The "Norm" observable was also removed from the list of observables in
the configuration files, since it is always measured anyway.