Add fTEDOPA benchmark script
Add a new script that may serve as a benchmark for our version of
fTEDOPA: it splits the environment in two parts, one with modes below
the chemical potential and the other above, and applies the standard
TEDOPA algorithm to both, resulting in two chains interacting with the
system. This is standard TEDOPA (or T-TEDOPA if the temperature is
positive) so the algorithm is already tested.