Add Mathematica script to compare correlation fn
Add a new Mathematica script which takes a "TEDOPA file", i.e. a
two-column file with couplings and frequencies, and computes the
correlation function (and the relative spectral density) associated to
the chain-mapped environment.
It doesn't work at the moment: the calculation of the correlation
function looks fine, but the derived spectral density function is wrong.