Add TrackerProjMPO type
Add a modified ProjMPO type that keeps track of the IDs of the bond
indices of the underlying state, and a modified position! method that
recomputes the relevant projection terms, so that the ProjMPO object
can change along with the state. This is useful e.g. when the state MPS
gets reorthogonalized, but the position of the ProjMPO is not moved,
in which case a standard ProjMPO would do nothing.