Use TrackerProjMPO to evaluate convergence measure
By using the new TrackerProjMPO type we can reuse it among different
calls of `bondconvergencemeasure`, instead of passing the Hamiltonian's
MPO and recalculating the ProjMPO from scratch each time.
Also, performing a QR decomposition instead of an SVD avoids the
(useless) step of getting S and V, which get multiplied in the end
anyway.