Use a single ProjMPO object in adaptive TDVP alg
With the flexibility provided by the new TrackerProjMPO object, it is
possible to initialize the projection object at the beginning of the
TDVP algorithm, and use it throughout the whole simulation, including
in the routine which increase the bond dimensions of the state.
This should save a lot of useless calculations.