Add function to force recomputation of projections
Add a function that recomputes the projections inside a ProjMPO object
on a given bond; we use this in specific points of the adaptation
routine, and we write explicitly on which sites we need to intervene,
instead of leaving all the guesswork to the ProjMPO object itself (which
was a failure). This function works with the AbstractProjMPO type so it
doesn't rely on the TrackerProjMPO features.