Cap bond dimension at lower values on MPS edges
When an upper limit is set on the bond dimensions of the evolved
MPS, the sites on the edge might never attain that limit, since their
bond dimensions are limited by their physical dimension: there is a
mathematical limit to their size.
Because of this, in the `adaptbonddimensions!` routine, the state on
the edges "do not know" that their bond dimension cannot grow higher
than this lower limit, and so the function tries anyway to see if it can
be enlarged. This is obviously just a waste of time; in fact, when the
MPS is reorthogonalized the bond dimension is brought back to the
mathematical limit.
As a solution, we use the minimum between the mathematical limit and the
"max_bond" parameter to decide whether to try to grow the bond
dimension each time.