Disable `ishermitian` kwarg in default solver
The `ishermitian` flag is set to true, but the default solver is used
with the GKSL equation too, and those operators are not Hermitian.
The `ishermitian` parameter must be passed in a different way if we want
that the solver takes it into consideration. For now, we just avoid
giving it a default value.