Add functions to build closure and chain ops
Add functions that simplify the creation of the (super-)operators
related to a spin chain and to a Markovian closure (both filled and
empty), so as to avoid duplicating the code among the simulation
scripts, and to make their creation easier. For now this is defined for
`vElectron` site types only.