Fix missing conjugation in `smart_contract`
The `smart_contract` function, used to compute the expectation value of an
operator on a MPS using the minimum number of sites needed to give a
correct answer, did not conjugate the "left" ITensor representing the
state, i.e. it computed psi^T * A * psi instead of psi^dagger * A * psi.
This corrects this bug which obviously produced nonsensical results in
the expectation values.