Use time-ind TDVP in relax phase, save MPS on disk
In the relaxation phase of the equilibration process, the
Hamiltonian/Lindbladian operator doesn't depend on time anymore, so the
time-dependent solver introduces an avoidable (big) overhead. By
stopping the time-dependent algorithm as soon as the ramp function
reaches one, we can save a significant amount of simulation time.
At the end of the simulation (and also at the end of the ramp, just in
case) we also save the MPS on disk, so that it can be used as an initial
state in other simulations.