Fix missing results in relaxation phase
In the SIAM equilibration scripts, the measurements in the relaxation
phase were never printed to the given output file. This was because at
the beginning of the relaxation phase the callback object storing the
measurements had its time equal to the final instant of the ramp-up
phase; however, the time began anew from zero in the relaxation phase,
so due to some conditionals in the measurement functions the
measurements were never performed.
Resetting the whole callback object before starting the relaxation phase
solves this.