Fix `growMPS` when initial dim. is not 1
The function (and its variants) assumed that the given MPS has link
dimensions all equal to 1. This fixes this behaviour, so that calling
`growMPS(v, N)` on `v` actually results in the MPS having link dimension
`N` on all bonds (as long as `N` is high enough that the operation makes
sense).