Do not write on HDF5 files if set to `/dev/null`
The `/dev/null` trick to avoid creating `.dat` files when printing the
measurements output does not work with the file storing the MPS, since
the `h5open` function does not seem to accept `/dev/null` as valid.
Instead, we check if the filename is `/dev/null` and we avoid opening
the file altogether.