Change `Measurement` into a simple Vector{Float64}
The `Measurement` type has always been an alias to a vector of vectors
of Float64, representing a set of measurements, one for each time step
and for each site of the system.
Since we never use this "global measurement" feature, but we use only
operators on specific sites, this double vector structure is
superfluous. A single vector is sufficient, so we reversed the
situation: now `Measurement`s are `Vector{Float64}`s, and if a callback
struct needs to have a Measurement on all sites, it will use a vector of
`Measurement`s.