Calculate TDVP1 exp. values with full MPO
Use a complete MPO (on all sites), instead of using the `smart_contract`
method which gave absurd results, to calculate expectation values of the
operators during the TDVP1 (non-vectorized) algorithm evolution. This is
probably a "dumb" way to do it, but at least it works. Memoizing the
step in which the MPO is created allows to save a large chunk of
computation time.