Fix `permute` and `check_hascommoninds` locations
The two names were directly prefixed with the name of the module they
live in (since they are probably not exported). The latter was recently
moved to the ITensorMPS subpackage, so the prefix had to be fixed.
By importing the names explicitly at the beginning of the file, it will
be easier in the future to change their prefix if it needs to be done.