Rescale initial state in MC Green's func script
Rescale the initial state so that its trace is 1. If it comes from a previous
TDVP evolution, its trace might have deviated a little from this value.
As long as it is a little different from 1, rescaling is fine and the
results of the Green's function will be more precise.