Fix output routine in `adjtdvp1vec!`
The part of the code which computes the expectation value of the evolved
operator and then outputs it was never completed, and as a result the
"measurement stride" parameter was neved considered. The script just
computed the value at each time step, regardless of it. This is now fixed,
in a way that mimics what the standard `tdvp1!` and `tdvp1vec!`
functions do with the help of the callback operators.