Add Green's func from pure state in Heisenberg p
Add a new script that calculates the (usual) part of the retarded Green's
function starting from a pure state, but evolving the d(t) operator in
the Heisenberg picture. In order to use the TDVP algorithm in the
Heisenberg picture, we need to transform the initial pure state in a
density matrix and vectorize it.