Add equilibration script with odd adj Lindblad op
Add a new version of the MC equilibration script that evolves the
initial state with the odd-parity adjoint Lindblad operator.
To compute Green's functions we need to evolve odd-grade operators, so
maybe the initial equilibrium state must also be prepared using the
odd-grade evolution operator (contrary to what we believed). We'll see
where this goes.