Use `chain_coefficients` for all coeff. files
Change the `"thermofield_coefficients"` and `"tedopa_coefficients"`
keys, both in the JSON files and in the simulation scripts, to
`"chain_coefficients"`, since the distinction was in name only anyway.
This makes the parameter files easier to write, since the user does not
have to guess whether to use "thermofield" or "tedopa" anymore.