Add first working script with superfermions
The superfermion formalism requires that we reformulate how expectation
values are computed, so we need a new kind of callback object to do
this, and a new method to increase the bond dimension too so that QN
indices are used correctly. We can use Fermion site types normally, but
it will be probably be necessary to write a certain amount of new helper
functions to construct the various terms of the Lindblad equation
easily.