Read parameters from JSON in superfermion script
Edit the superfermion script (without Markovian closure) so that it can
accept parameters from the usual JSON file as well as from the command
line.  In the future we may use this input strategy as the default one,
since it adds more flexibility to the currently used one with JSON only.