Move TimeEvoVecMPS into separate package
Detach the TimeEvoVecMPS package and create a new separate Julia package
from it: it is called MPSTimeEvolution and can be found in the
TensorNetworkSimulations register.
All functions from TimeEvoVecMPS were removed from this project except
`load_pars`, which is outside the scope of the package, so it was moved
instead in a file included by all simulation scripts.