First test of "output packing"
Write some more code after the end of the simulation in
`superfermions/test_1env.jl` to gather all the results and the input
parameters in a single file, so that it's very easy to keep track
of which parameters were used to obtain those results.
It will be later made into a separate function that can be shared among
all the simulation scripts.