Add function to reverse ModeChain objects
Add a function that reverses the order of the components of a
ModeChain object. It's useful when we want to join the "system"
chain and the two ones coming from the two environments: we reverse one
of the environments and join the three chains together by calling
something like `join(reverse(cL), join(cS, cR))`. Without
reversing one of them it would be impossible to use the `join` function
properly like this. We cannot just sum `join(cS, cL)` and `join(cS, cR)`
because that would make the system appear twice in the Hamiltonian.