Generalise `join` to non-adjacent `ModeChain`s
Generalise the `join` method for `ModeChain`s so that the two chains do
not need to be completely separated, i.e. `c1[end] < c2[begin]`, but
they can also overlap, meaning that we can join e.g. chains on sites
`[2, 4, 6]` and `[3, 5, 7]`. The only requirement is that their
ranges do not intersect.