Create core package `OpenSystemsChainMapping`
Move core methods such as utility functions, argument parsers,
`ModeChain` methods, and some definition of physical systems into a
Julia package so that the code can be better organised and "end" scripts
can load the relevant code with just a `using OpenSystemsChainMapping`
line instead of not-so-obvious `include` directives.
This way it's also easier to reproduce the working environment in new
installations of the repository.