Fix CSV columns not read as simple vectors
If Julia is run with multithreading active and the CSV file is large
(> 5000 cells), `CSV.read` uses multiple threads to read it, and
its columns will be read as `SentinelArrays.ChainedVector{T, Vector{T}}`
instead of `Vector{T}`.
Use `collect` in the `pack!` function to transform the former into the
latter, in case this happens, since `HDF5.write` accepts only simple
vectors, otherwise it throws an error and doesn't write the data.