Remove compat restrictions on OpenSystemsChainMapping
Remove all compatibility restrictions on dependencies but the one on
MarkovianClosure, which is the only sensible one: we actually need a
sufficiently recent version so that the MC coefficients are "baked in"
the package. The other restrictions appeared in the Project file by
mistake and do not serve any purpose, if anything they prevent some
needed updates from being downloaded.