Promote type correctly for evolved corr. matrices
Sometimes the initial correlation matrix has `Float64` as element type but
the generator matrix has `ComplexF64`, so if we use `eltype(initmatrix)`
to create the array for the evolved matrices Julia complains that it
cannot represent a complex number as a `Float64`.
Instead, use Julia's builtin promotion rules to create the array for the
evolved matrices with the appropriate supertype of the initial matrix
and the generator matrix types.