Generalise type in fixed-step corr. mat. evolution
Compute the matrix exponential via diagonalisation instead of using
Julia's built-in `exp`, which doesn't work e.g. when the element type of
the matrix is `BigFloat`. The `eigen` function and the exponential of a
single number are defined for `BigFloat`s, instead, so a diagonal matrix
of `BigFloat` numbers constructed with the `Diagonal` type can be safely
exponentiated.