Use natural indexing for corr. fn. submatrix
Instead of looping through a list of indices, pass a single, generic
`inds...` to the `getindex` function that can be used to return a
submatrix of the correlation function matrix. This is a more intuitive
and flexible syntax, and the return values are stored in the same way as
in the original matrix of correlation functions.
Add progress bars.